+++ src/test/resources/case/control_flow/if/main.txt
@@ -1,0 +1,12 @@
+var x: i32 = if 2 > 1
+    0
+else
+    5
+System.print(x)
+
+var y: i32 = if 5 < 2 {
+    true //Condition is constant false, so it doesn't bother type-checking this branch
+} else {
+    10
+}
+System.print(y)
