+++ src/test/resources/case/control_flow/while/main.txt
@@ -1,0 +1,4 @@
+var x: i32 = 10
+while x > 0 {
+    System.print(x = x - 1);
+}
