+++ src/test/resources/case/uncategorized/2classes/main.txt
@@ -1,0 +1,11 @@
+class Thing {
+    pub fn new() super()
+    pub fn get5(): i32
+        1 + 1 + 1 + 1 + 1
+}
+class Thang {
+    pub fn new() super()
+    pub fn get5Again(thing: Thing): i32
+        thing.get5()
+}
+System.print(new Thang().get5Again(new Thing()))
