+++ src/test/resources/case/uncategorized/i2l/main.txt
@@ -1,0 +1,3 @@
+var x: i32 = 10
+var y: i64 = System.i2l(x)
+Test.assertEquals(10i64, y)
