+++ src/test/resources/case/uncategorized/int_prints/main.txt
@@ -1,0 +1,27 @@
+var x: i32 = 2147483647;
+x = x + 2;
+var y: u32 = 2147483647;
+y = y + 2;
+
+var z: i64 = 9223372036854775807;
+z = z + 2;
+var w: u64 = 9223372036854775807;
+w = w + 2;
+System.print(x)
+System.print(y)
+System.print(z)
+System.print(w)
+
+var a: u32 = 4294967295
+a = a + 1
+System.print(a)
+a = a - 1
+System.print(a)
+
+var b: i32 = -1
+var c: u32 = 0
+c = c - 1;
+System.print(b * b)
+System.print(c * c)
+System.print(b / 10)
+System.print(c / 10)
